pyproject.toml
[project]
name = "cbning"
version = "0.1.0"
description = "Causal Bayesian Network Builder"
authors = [{ name = "Orpheus Lummis", email = "o@orpheuslummis.info" }]
dependencies = [
    "gradio>=5.0.1",
    "litellm>=1.49.0",
    "networkx>=3.3",
    "plotly>=5.24.1",
    "click>=8.1.3",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.9.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
cbning = "cbning.app:launch_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cbning"]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "flake8"]
